// lib/services/api/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  AmountHumanReadable: { input: string; output: string };
  BigDecimal: { input: string; output: string };
  BigInt: { input: string; output: string };
  Bytes: { input: string; output: string };
  Date: { input: any; output: any };
  GqlBigNumber: { input: string; output: string };
  JSON: { input: any; output: any };
};

/** The review data for the ERC4626 token */
export type Erc4626ReviewData = {
  __typename: "Erc4626ReviewData";
  /** The filename of the review of the ERC4626 */
  reviewFile: Scalars["String"]["output"];
  /** A summary of the ERC4626 review, usually just says safe or unsafe */
  summary: Scalars["String"]["output"];
  /** Warnings associated with the ERC4626 */
  warnings: Array<Scalars["String"]["output"]>;
};
// ...
export enum GqlChain {
  Base = "BASE",
  Mainnet = "MAINNET",
  Sepolia = "SEPOLIA",
  Sonic = "SONIC",
  Arbitrum = "ARBITRUM",
}
// ...
export type GqlPriceRateProviderData = {
  __typename: "GqlPriceRateProviderData";
  /** The address of the price rate provider */
  address: Scalars["String"]["output"];
  /** The factory used to create the price rate provider, if applicable */
  factory?: Maybe<Scalars["String"]["output"]>;
  /** The name of the price rate provider */
  name?: Maybe<Scalars["String"]["output"]>;
  /** The filename of the review of the price rate provider */
  reviewFile?: Maybe<Scalars["String"]["output"]>;
  /** Indicates if the price rate provider has been reviewed */
  reviewed: Scalars["Boolean"]["output"];
  /** A summary of the price rate provider, usually just says safe or unsafe */
  summary?: Maybe<Scalars["String"]["output"]>;
  /** Upgradeable components of the price rate provider */
  upgradeableComponents?: Maybe<
    Array<Maybe<GqlPriceRateProviderUpgradeableComponent>>
  >;
  /** Warnings associated with the price rate provider */
  warnings?: Maybe<Array<Scalars["String"]["output"]>>;
};

/** Represents an upgradeable component of a price rate provider */
export type GqlPriceRateProviderUpgradeableComponent = {
  __typename: "GqlPriceRateProviderUpgradeableComponent";
  /** The entry point / proxy of the upgradeable component */
  entryPoint: Scalars["String"]["output"];
  /** Indicates if the implementation of the component has been reviewed */
  implementationReviewed: Scalars["String"]["output"];
};
// ...
export type GqlToken = {
  __typename: "GqlToken";
  /** The address of the token */
  address: Scalars["String"]["output"];
  /** The chain of the token */
  chain: GqlChain;
  /** The chain ID of the token */
  chainId: Scalars["Int"]["output"];
  /** The coingecko ID for this token, if present */
  coingeckoId?: Maybe<Scalars["String"]["output"]>;
  /** The number of decimal places for the token */
  decimals: Scalars["Int"]["output"];
  /** The description of the token */
  description?: Maybe<Scalars["String"]["output"]>;
  /** The Discord URL of the token */
  discordUrl?: Maybe<Scalars["String"]["output"]>;
  /** The ERC4626 review data for the token */
  erc4626ReviewData?: Maybe<Erc4626ReviewData>;
  /** If it is an ERC4626 token, this defines whether we allow it to use the buffer for pool operations. */
  isBufferAllowed: Scalars["Boolean"]["output"];
  /** Whether the token is considered an ERC4626 token. */
  isErc4626: Scalars["Boolean"]["output"];
  /** The logo URI of the token */
  logoURI?: Maybe<Scalars["String"]["output"]>;
  /** The name of the token */
  name: Scalars["String"]["output"];
  /** The rate provider data for the token */
  priceRateProviderData?: Maybe<GqlPriceRateProviderData>;
  /** The priority of the token, can be used for sorting. */
  priority: Scalars["Int"]["output"];
  /**
   * The rate provider data for the token
   * @deprecated Use priceRateProviderData instead
   */
  rateProviderData?: Maybe<GqlPriceRateProviderData>;
  /** The symbol of the token */
  symbol: Scalars["String"]["output"];
  /** The Telegram URL of the token */
  telegramUrl?: Maybe<Scalars["String"]["output"]>;
  /** Indicates if the token is tradable */
  tradable: Scalars["Boolean"]["output"];
  /** The Twitter username of the token */
  twitterUsername?: Maybe<Scalars["String"]["output"]>;
  /** The ERC4626 underlying token address, if applicable. */
  underlyingTokenAddress?: Maybe<Scalars["String"]["output"]>;
  /** The website URL of the token */
  websiteUrl?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type GqlTokenPrice = {
  __typename: "GqlTokenPrice";
  address: Scalars["String"]["output"];
  chain: GqlChain;
  price: Scalars["Float"]["output"];
  updatedAt: Scalars["Int"]["output"];
  updatedBy?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type GetTokensQueryVariables = Exact<{
  chains: Array<GqlChain> | GqlChain;
}>;

export type GetTokensQuery = {
  __typename: "Query";
  tokens: Array<{
    __typename: "GqlToken";
    address: string;
    name: string;
    symbol: string;
    decimals: number;
    chain: GqlChain;
    chainId: number;
    logoURI?: string | null;
    priority: number;
    tradable: boolean;
    isErc4626: boolean;
    isBufferAllowed: boolean;
    coingeckoId?: string | null;
  }>;
};

export type GetTokenPricesQueryVariables = Exact<{
  chains: Array<GqlChain> | GqlChain;
}>;

export type GetTokenPricesQuery = {
  __typename: "Query";
  tokenPrices: Array<{
    __typename: "GqlTokenPrice";
    price: number;
    address: string;
    chain: GqlChain;
    updatedAt: number;
  }>;
};
// ...
export const GetTokensDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "GetTokens" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "chains" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "ListType",
              type: {
                kind: "NonNullType",
                type: {
                  kind: "NamedType",
                  name: { kind: "Name", value: "GqlChain" },
                },
              },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            alias: { kind: "Name", value: "tokens" },
            name: { kind: "Name", value: "tokenGetTokens" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "chains" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "chains" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "address" } },
                { kind: "Field", name: { kind: "Name", value: "name" } },
                { kind: "Field", name: { kind: "Name", value: "symbol" } },
                { kind: "Field", name: { kind: "Name", value: "decimals" } },
                { kind: "Field", name: { kind: "Name", value: "chain" } },
                { kind: "Field", name: { kind: "Name", value: "chainId" } },
                { kind: "Field", name: { kind: "Name", value: "logoURI" } },
                { kind: "Field", name: { kind: "Name", value: "priority" } },
                { kind: "Field", name: { kind: "Name", value: "tradable" } },
                { kind: "Field", name: { kind: "Name", value: "isErc4626" } },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "isBufferAllowed" },
                },
                { kind: "Field", name: { kind: "Name", value: "coingeckoId" } },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<GetTokensQuery, GetTokensQueryVariables>;
export const GetTokenPricesDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "GetTokenPrices" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "chains" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "ListType",
              type: {
                kind: "NonNullType",
                type: {
                  kind: "NamedType",
                  name: { kind: "Name", value: "GqlChain" },
                },
              },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            alias: { kind: "Name", value: "tokenPrices" },
            name: { kind: "Name", value: "tokenGetCurrentPrices" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "chains" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "chains" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "price" } },
                { kind: "Field", name: { kind: "Name", value: "address" } },
                { kind: "Field", name: { kind: "Name", value: "chain" } },
                { kind: "Field", name: { kind: "Name", value: "updatedAt" } },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<GetTokenPricesQuery, GetTokenPricesQueryVariables>;